+++ internal/servers/destination/v0/destinations.go
@@ -141,6 +141,7 @@ func (s *Server) Write2(msg pb.Destination_Write2Server) error {
 		select {
 		case resources <- convertedResource:
 		case <-ctx.Done():
+			convertedResource.Release()
 			close(resources)
 			if err := eg.Wait(); err != nil {
 				return status.Errorf(codes.Internal, "Context done: %v and failed to wait for plugin: %v", ctx.Err(), err)