+++ plugins/destination/plugin_testing_overwrite_delete_stale.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/apache/arrow/go/v12/arrow"
 	"github.com/apache/arrow/go/v12/arrow/array"
-	"github.com/apache/arrow/go/v12/arrow/memory"
 	"github.com/cloudquery/plugin-sdk/v2/specs"
 	"github.com/cloudquery/plugin-sdk/v2/testdata"
 	"github.com/cloudquery/plugin-sdk/v2/types"
@@ -15,7 +14,7 @@ import (
 	"github.com/rs/zerolog"
 )
 
-func (*PluginTestSuite) destinationPluginTestWriteOverwriteDeleteStale(ctx context.Context, mem memory.Allocator, p *Plugin, logger zerolog.Logger, spec specs.Destination) error {
+func (*PluginTestSuite) destinationPluginTestWriteOverwriteDeleteStale(ctx context.Context, p *Plugin, logger zerolog.Logger, spec specs.Destination) error {
 	spec.WriteMode = specs.WriteModeOverwriteDeleteStale
 	if err := p.Init(ctx, logger, spec); err != nil {
 		return fmt.Errorf("failed to init plugin: %w", err)
@@ -44,18 +43,10 @@ func (*PluginTestSuite) destinationPluginTestWriteOverwriteDeleteStale(ctx conte
 		SyncTime:   syncTime,
 		MaxRows:    2,
 	}
-	resources := testdata.GenTestData(mem, table.ToArrowSchema(), opts)
-	incResources := testdata.GenTestData(mem, incTable.ToArrowSchema(), opts)
+	resources := testdata.GenTestData(table.ToArrowSchema(), opts)
+	incResources := testdata.GenTestData(incTable.ToArrowSchema(), opts)
 	allResources := resources
 	allResources = append(allResources, incResources...)
-	for _, r := range allResources {
-		r.Retain()
-	}
-	defer func() {
-		for _, r := range allResources {
-			r.Release()
-		}
-	}()
 	if err := p.writeAll(ctx, sourceSpec, syncTime, allResources); err != nil {
 		return fmt.Errorf("failed to write all: %w", err)
 	}
@@ -98,9 +89,7 @@ func (*PluginTestSuite) destinationPluginTestWriteOverwriteDeleteStale(ctx conte
 		StableUUID: *u,
 		MaxRows:    1,
 	}
-	updatedResources := testdata.GenTestData(mem, table.ToArrowSchema(), opts)[0]
-	updatedResources.Retain()
-	defer updatedResources.Release()
+	updatedResources := testdata.GenTestData(table.ToArrowSchema(), opts)[0]
 
 	if err := p.writeOne(ctx, sourceSpec, secondSyncTime, updatedResources); err != nil {
 		return fmt.Errorf("failed to write one second time: %w", err)